requirements.txt
attr==0.3.1
attrs==17.2.0
Augmentor==0.1.10
backports-abc==0.5
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.48.0
certifi==2017.11.5
config==0.3.9
decorator==4.1.2
funcsigs==1.0.2
future==0.16.0
glob2==0.6
html5lib==0.9999999
IDLSave==1.0.0
imgaug==0.2.5
Keras==2.0.6
Markdown==2.6.10
matplotlib==1.3.1
madmom==0.15.1
mido==1.2.8
mock==2.0.0
music21==4.1.0
networkx==2.0
nose==1.3.7
numpy==1.13.1
olefile==0.44
opencv-python==3.3.0.9
pandas==0.20.3
pbr==3.1.1
Pillow==4.3.0
protobuf==3.5.1
psutil
pyparsing==2.2.0
python-dateutil==2.6.1
tables
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
retrying==1.3.3
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==0.19.1
singledispatch==3.4.0.3
six==1.11.0
tensorflow-gpu
Theano==1.0.1
tornado==4.5.2
tqdm==4.19.5
Werkzeug==0.13
tables